refactored home page to be server side and load popular songs SSR
